backend-scrapper/requirements.txt
requests>=2.32.0
supabase>=2.13.0
python-dotenv>=1.0.1
tenacity>=9.0.0
click>=8.1.7
python-dateutil>=2.9.0
